add request

+++ services/storage.service.js
@@ -1,54 +1,62 @@
+import { homedir } from "os";
+import { join } from "path";
+import { promises } from "fs";
+import { printError, printSuccess, printInfo } from "./log.service.js";
 
-import { homedir } from 'os'
-import { join } from 'path'
-import { promises } from 'fs'
-import { printError, printSuccess, printInfo} from './log.service.js'
+const filePath = join(homedir(), "weather-app-data.json");
 
-const filePath = join(homedir(), 'weather-app-data.json')
+const TOKEN_DICTIONARY = {
+    token: "token",
+    city: "city",
+}
 
 const saveKeyValue = async (key, value) => {
-    let data = {}
+    let data = {};
     if (await isExist(filePath)) {
-        const file = await promises.readFile(filePath)
-        data = JSON.parse(file)
+        const file = await promises.readFile(filePath);
+        data = JSON.parse(file);
     }
-    data[key] = value
-    await promises.writeFile(filePath, JSON.stringify(data))
-}
+    data[key] = value;
+    await promises.writeFile(filePath, JSON.stringify(data));
+};
 
 const getKeyValue = async (key) => {
     if (await isExist(filePath)) {
-        const file = await promises.readFile(filePath)
-        const data = JSON.parse(file)
-        return data[key]
+        const file = await promises.readFile(filePath);
+        const data = JSON.parse(file);
+        return data[key];
     }
-    return undefined
-}
+    return undefined;
+};
 
 const isExist = async (path) => {
     try {
-        await promises.stat(path)
-        return true
+        await promises.stat(path);
+        return true;
     } catch (e) {
-        return false
+        return false;
     }
-}
+};
 
 const saveToken = async (token) => {
+    if (token.length < 32) {
+        printError(`The token must be at least 32 characters, now ${token.length}`);
+        return;
+    }
+
     try {
-        await saveKeyValue('token', token)
-        printSuccess('Token saved')
+        await saveKeyValue(TOKEN_DICTIONARY.token, token);
+        printSuccess("Token saved");
     } catch (error) {
         if (error instanceof SyntaxError) {
-            promises.unlink(filePath)
-            printInfo()
-            saveToken(token)
+            promises.unlink(filePath);
+            printInfo();
+            saveToken(token);
         } else {
-            console.log(error)
-            printError(error.message)
+            console.log(error);
+            printError(error.message);
         }
-        
     }
-}
+};
 
-export { saveKeyValue, getKeyValue, saveToken }
+export { saveKeyValue, getKeyValue, saveToken, TOKEN_DICTIONARY };
